Xmin horizon check: Show TXID with separate epoch, and assigned at time
Since the statistics views in Postgres utilize the 32-bit version of
transaction IDs, but we track 64-bit internally, it can be confusing
to show the raw 64-bit value (since you can't visually compare it to
queries you run on the database). To improve, instead format it with
a separate epoch, separated by a colon, like we do for freezing stats.

Additionally show the specific assigned at timestamp inline, to help
identify the problem origin better, e.g. to use for comparison with
xact_start in pg_stat_activity, when xmin horizon is held back by a
long-running transaction.

In passing introduce a component to handle xmin information consistently,
and fix typing for "heldBackInfo", which was incorrectly marked as a raw
number (its an object).